requirements.txt
torch
torchvision
streamlit
streamlit-drawable-canvas
Pillow
